back/restApiMountainPeaks/requirements.txt
Django==3.2.6
djangorestframework==3.13.1  
psycopg2>=2.8
django-cors-headers==3.11.0